FE y BE funcionales, creación de usuarios correcta, delete en progreso

+++ gestion-gastos-front/src/components/Main.jsx
@@ -3,25 +3,39 @@ import { useAuth } from "../Contexts/FBauthContext/index.jsx";
 import { AuthContext } from "../Contexts/FBauthContext/index.jsx";
 import { fbDeleteUser } from "../Firebase/auth.js";
 import { getAuth } from "firebase/auth";
-const auth = getAuth();
+
 const userDeleteManager = async (user) => {
   try {
     // Esto borra el usuario de Firebase
-    fbDeleteUser(auth.user, auth.user.email, prompt("confirme su contraseña"));
+    const auth = getAuth();
+    const user = auth.currentUser;
+    fbDeleteUser(user, user.email, prompt("confirme su contraseña"));
 
     //Borra el usuario de la base de datos
     //Obtiene el token de identidad para autenticar al usuario,
     //  refrescando el id para evitar su vencimiento durante el proceso.
+    if (!user) {
+      console.error("No user is currently signed in.");
+      return;
+    }
     const token = await auth.currentUser.getIdToken(true);
-    fetch(`http://localhost:3001/api/user/`, {
+    const response = await fetch(`http://localhost:3001/api/user/${user.uid}`, {
       method: "DELETE",
       headers: {
         "Content-Type": "application/json",
-        authorization: `bearer ${token}`,
+        Authorization: `bearer ${token}`,
       },
       body: JSON.stringify({ email: user.email }),
     });
-    console.log("usuario eliminado de la base de datos");
+    if (response.ok) {
+      console.log("usuario eliminado de la base de datos");
+    } else {
+      const errorData = await response.json();
+      console.error(
+        "FE: Error eliminando usuario de la base de datos:",
+        errorData
+      );
+    }
   } catch (err) {
     console.error("FE: Error deleting user:", err);
   }
